refactor: change components directory related to options page
