+++ src/options/index.tsx
@@ -12,13 +12,13 @@ import {
 } from "@mui/material";
 import { blue, grey } from "@mui/material/colors";
 import React, { useEffect, useState } from "react";
-import { AdvancedOptions } from "./components/AdvancedOptions";
-import { BasicOptions } from "./components/BasicOptions";
-import { CustomCSS } from "./components/CustomCSS";
-import { DataOperation } from "./components/DataOperation";
-import { Information } from "./components/Information";
-import { LayoutOptions } from "./components/LayoutOptions";
-import { WidgetOptions } from "./components/WidgetOptions";
+import { AdvancedOptions } from "~/options/components/pages/AdvancedOptions";
+import { BasicOptions } from "~/options/components/pages/BasicOptions";
+import { CustomCSS } from "~/options/components/pages/CustomCSS";
+import { DataOperation } from "~/options/components/pages/DataOperation";
+import { Information } from "~/options/components/pages/Information";
+import { LayoutOptions } from "~/options/components/pages/LayoutOptions";
+import { WidgetOptions } from "~/options/components/pages/WidgetOptions";
 import theme from "~/theme";
 import { defaultSaves } from "~settings";
 import type { Saves } from "~settings";
